Implement traps as a custom instruction, include a custom OS page 0
Implementing traps as a custom instruction is much nicer than trapping on
memory reads, since it's a sure sign that the given code is being executed
rather than just being read by something. To support it, we need to map
something into the page 0 region of memory, which is a little OS image that
currently just contains traps.