Turn off regex logging support
regex is a pretty heavy dependency, so dropping it from default builds improves
build times pretty significantly while it can still be turned on explicitly.

In a quick test, turning off regex on my 4-CPU system saves a minute of CPU
time and 20 seconds of real time on release builds.