Schedule interrupts based on emulated, not wall time
Wall time kind of works as long as the system is running at the intended
speed, but we need to track emulated time instead to make it work correctly
when running at any other speed.

Also reduce the timer1 default frequency to 118 Hz, in line with what wikiti
says. The 140Hz value seems to come from 83pa28d, which is probably less
accurate.