Implement a "linker" around spasm for building the OS
We can declare what page a binary is and the symbols it imports/exports;
the "linker" runs spasm and pulls label values out of the emitted
symbol table, feeding them in via the command line to files that import
them.

This is kind of like a smarter version of make, in that it can determine
the order in which the specified files must be built by topologically
sorting the dependency graph. A real linker might be better in the long
term.