Synchronize debuggers, support async break and ^C to trigger it
On connection, the debugger synchronizes with the emulator to flush any
excess responses that may go with a client that disconnected. Without this,
it's possible to get a client behaving strangely because it's sending commands
but getting unrelated responses.

For convenience, handle Ctrl+C in the client (though on Windows it's kind of
unreliable) and send an interrupt. The emulator can also be paused
interactively; a good alternative on Windows.

Async break fits user-triggered break into the Wait command, allowing all CPU
runs to work in the same way. The Interrupt command never receives a response,
which allows clients to send them at any time and end a Wait- meaning a client
can of its own accord end a Wait in the same way as it would normally.